requirements.txt
Django==4.2.4
openai==0.27.8
python-dotenv==1.0.0
